requirements.txt
torch>=2.2.0
transformers>=4.57.0
accelerate>=0.33.0
pillow>=10.0.0
decord>=0.6.0
opencv-python>=4.9.0
sentencepiece
safetensors
requests
tqdm
torchvision==0.23.0
vllm==0.11.0
qwen-vl-utils==0.0.14
scikit-image
peft
trl
bitsandbytes
wandb
ray[tune]
accelerate